+++ src/create-cached-async-fnc.ts
@@ -1,4 +1,3 @@
-import { cloneDeep } from "./clone-deep";
 import { generateId } from "./generate-id";
 import { toConsole } from "./to-console";
 
@@ -9,6 +8,7 @@ type Config = {
 /**
  * @description
  * This function creates a cachedAsyncFnc instance which you can use to execute the resolverFunction with different arguments.
+ * @deprecated This function will be removed in v3 please migrate to the CachedFunction class
  * @example
  * const cacheFnc = createCachedAsyncFnc(async (userId: string) => {
  *   const user = await fetchUser(userId);