+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0 - 2023-07-05
+
+### Added
+
+- Implemented a new `CachedFunction` class with better types and simpler code
+
+### Deprecated
+
+- The old `createCachedAsyncFnc()` apis is now deprecated please migrate to the CachedFunction class
+
+### Removed
+
+- Drop support for Node v14
+
 ## 1.0.2 - 2022-06-23
 
 ### Changed